Refaktorerer hvordan k9beskjed blir bygget opp
